Move query endpoints to the underlays API section.
